created separate view files
